added abstract class, black formatted, fixed all linting errors, updated README for test running
